Amazing performance improvements in VBAP mode.
This is done by short-circuiting the spatialization process when is makes
sense. This should not change much when everything is throwing sound
everywhere, but this is a rare case since the speakers are organized in
triplets.

Note : while this does not solve issue #209, it is an acceptable
quickfix since the extra channels will never send any data.

This is directly related to issue #192.