Add the option to send extra data to a standalone SpeakerView instead of redirecting the data flow (#513)
The option for SpeakerView data used to redirect the UDP messages to another address/port. Now setting these options send the data to another target using another socket.

This enables workflows like having the spatialization server in the venue's server but sending the UDP to the artist's computer so that the venue's technician and the artist can monitor at the same time or even receiving the spatialization data to something like touchdesigner to enable spatialization related visuals while still monitoring with SpeakerView.