Fix speaker ordering ignoring the speaker setup window's order for speaker groups (#529)
Before this, when you moved a speaker group around in the speaker setup window, the vue meter would not honour the order of the speakers in the groups.

This also fixes an issue where the save dialog window would not be shown if a group was only moved in an existing speaker setup.