Fix ordering bug when adding speakers and groups to speaker setups that already contains a speaker group  (#533)
Fix a bug where the speaker setup edition window got confused on orderings when adding a speaker or speaker group in a speaker setup that already contained a group.

After this, added speakers and speaker groups should be right after the last selected speaker/group and the order displayed in the vue meter should be sync'ed with the order displayed in the speaker setup window.