Add a "use multicore DSP" button that enables parallel processing in VBAP, MPAB and hybrid mode (#544)
This builds on the work of @vberthiaume to polish fork_union integration in the spatialization algorithms.

Since parallel processing increases resource usage across the board and it may have different latency characteristic due to using a sleep() in the threadpool (necessary to not use 100%CPU at all time), I made it opt in with a new button in the interface.

Speedups can go from about 20% up to 100% depending on the use case. In my empirical testing I have been able to run 256 sources with active audio, high elevation and amplitude spans and constant motion through a speaker setup with 60 speakers with the MBAP algorithm without missing samples. The same setup did not work without the parallel processing.

* fork_union again

* numa

* more numa but why macos ?

* disable numa

* really sleep 1us in mbap

* switch to a debugged fork of FU 2.2.9

* some work on runtime switchable fork_union

* properly handles hybrid algorithm

* parallelize button somewhat works

* better layout for the multicore button + save and load behaviour fixes

* remove fork union macros and update algogris

* update algogris

* update algogris + suggested radio button changes

* use juce::Component::SafePointer instead of raw pointers

* add a TODO

* pass the boolean flag to initButton

* formatting